% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_followers}
\alias{get_followers}
\title{gets the current followers of a given Twitch user as identified by its user-id}
\usage{
get_followers(client_id, user_id, max_requests = 10)
}
\arguments{
\item{client_id}{A Twitch API client id}

\item{user_id}{A Twitch user id}

\item{max_requests}{The max number of requests you want to send (each request returns a maxiumum of 100 followers)}
}
\value{
A dataframe containing the current information for followers of a Twitch user.
}
\description{
gets the current followers of a given Twitch user as identified by its user-id
}
