% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_games}
\alias{get_games}
\title{gets the current information for a list of games featured on Twitch identified by game-ids}
\usage{
get_games(client_id, game_ids)
}
\arguments{
\item{client_id}{A Twitch API client id}

\item{game_ids}{A vector of game ids}
}
\value{
A dataframe containing the current information for a list of game titles on Twitch.
}
\description{
gets the current information for a list of games featured on Twitch identified by game-ids
}
