% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_streams}
\alias{get_streams}
\title{gets the top current streams on Twitch by viewers}
\usage{
get_streams(client_id, pages = 10)
}
\arguments{
\item{client_id}{A Twitch API client id}

\item{pages}{The number of pages of streams returned, each page has around 20 observations}
}
\value{
A dataframe containing the top current streams on Twitch by viewers
}
\description{
gets the top current streams on Twitch by viewers
}
