% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_users}
\alias{get_users}
\title{gets the current information for a list of Twitch users identified by user-ids}
\usage{
get_users(client_id, user_ids)
}
\arguments{
\item{client_id}{A Twitch API client id}

\item{user_ids}{A vector of user ids}
}
\value{
A dataframe containing the current information for a list of Twitch users.
}
\description{
gets the current information for a list of Twitch users identified by user-ids
}
