% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{twitch_request_into_df}
\alias{twitch_request_into_df}
\title{Helper function to turn requests from the Twitch API into dataframes}
\usage{
twitch_request_into_df(request)
}
\arguments{
\item{request}{A request returned from a call to the Twitch API}
}
\value{
A dataframe containing the data returned from a call to the Twitch API
}
\description{
Helper function to turn requests from the Twitch API into dataframes
}
